feat(routing): merge /app into /, move /app/reservations to /reservations
